pyproject.toml
[project]
name = "jaudible"
version = "0.0.1"
description = "Generate audio books from text documents"
authors = [
    { name="James Saryerwinnie", email="js@jamesls.com" }
]
license = "Apache-2.0"
readme = "README.md"


classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "Operating System :: OS Independent",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.12",
  "Topic :: Software Development :: Libraries :: Python Modules",
  "Typing :: Typed",
]
requires-python = ">=3.12.0"

dependencies = [
  "typer>=0.12.3,<0.13.0",
  "boto3>=1.34.133,<2.0.0",
  "botocore-stubs>=1.34.132,<2.0.0",
  "boto3-stubs[polly,s3]>=1.34.133,<2.0.0",
]

[tool.uv]
dev-dependencies = [
  "pre-commit",
  "pytest",
  "pytest-cov",
  "mkdocstrings",
  "mkdocs-material",
  "pymdown-extensions",
  "pytest-github-actions-annotate-failures",
  "jmeslog",
  "poethepoet",
  "ruff",
  "pyright",
  "pygments",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/jaudible"]


[tool.ruff]
line-length = 79

[tool.ruff.format]
quote-style = "preserve"
preview = true

[tool.pytest.ini_options]
addopts = """\
    --cov jaudible \
    --cov-report term-missing \
    --no-cov-on-fail \
"""

[tool.coverage.run]
branch = true
source = ["src/"]

[tool.coverage.report]
exclude_lines = [
    'if TYPE_CHECKING:',
    'pragma: no cover',
    'raise NotImplementedError',
]

[tool.pyright]
include = ["src"]
exclude = ["**/__pycache__"]

reportMissingImports = true
reportMissingTypeStubs = false
reportTypedDictNotRequiredAccess = false

pythonVersion = "3.12"
pythonPlatform = "All"

[tool.poe.tasks]
pre-commit = "poetry run pre-commit run --all-files"
check = [
{ cmd = "uv run ruff check ." },
{ cmd = "uv run ruff format --diff ." },
{ cmd = "uv run pyright" },
]
test = "uv run pytest tests/"
stage-release = "python3 ./scripts/stage-release --output staged-release.zip"
format-code = [
  { cmd = "uv run ruff format ." },
]
prcheck = ["check", "test", "pre-commit"]

# Create a jmes-tts CLI command
[project.scripts]
jmes-tts = "jaudible.cli:app"
